Extracting further in to a TooManyOfItemInBasketException
